Make torch._guards, shuffle structures around for migration
[ghstack-poisoned]
