+++ torch/_dynamo/source.py
@@ -4,7 +4,7 @@ from typing import Any
 
 from . import utils
 from .bytecode_transformation import create_instruction
-from .guards import Guard, GuardSource
+from torch._guards import Guard, GuardSource
 from .utils import rename_implicit
 
 _GUARD_SOURCE_NN_MODULE = {
