+++ torch/_dynamo/variables/misc.py
@@ -8,7 +8,7 @@ import torch._C
 from .. import config, variables
 from ..bytecode_transformation import create_instruction
 from ..exc import unimplemented
-from ..guards import Guard, GuardBuilder, GuardSource
+from ..guards import GuardBuilder
 from ..source import AttrSource
 from ..utils import identity, proxy_args_kwargs
 from .base import VariableTracker
@@ -18,6 +18,7 @@ from .functions import (
     WrappedUserFunctionVariable,
     WrappedUserMethodVariable,
 )
+from torch._guards import Guard, GuardSource
 
 
 class SuperVariable(VariableTracker):
